Copter: move surface tracking implementation to surface_tracking.cpp
